git-svn-id: svn://svn.dre.vanderbilt.edu/DOC/Middleware/trunk/ACE@80826 ae88bc3d-4319-0410-8dbf-d08b4c9d3795

+++ ace/Log_Msg_IPC.h
@@ -57,9 +57,17 @@ public:
 
   /// Open a new connection
   virtual int open (const ACE_TCHAR *logger_key);
+
+  /**
+   * Reset the backend.  When changing the logging destination the
+   * backend may need to properly disconnect from the remote logging
+   * daemon and reclaim some local resources.  But we try to reduce
+   * the number of local allocations/deallocations.
+   */
   virtual int reset (void);
+
   virtual int close (void);
-  virtual int log (ACE_Log_Record &log_record);
+  virtual ssize_t log (ACE_Log_Record &log_record);
 
 private:
   ACE_LOG_MSG_IPC_STREAM message_queue_;
